работающие проработанные тесты (и валидация, и КБЖУ)

+++ Test.Core/Helpers/ProductHelper.cs
@@ -163,14 +163,14 @@ public static class ProductHelper
         Protein,
         Sugar
     ];
-    public static (List<Ingredient>, List<Product>) BuildIngredientsAndProductsByIdsAndAmounts(int[] productIdsAndAmounts)
+    public static (List<Ingredient>, List<Product>) BuildIngredientsAndProductsByIdsAndAmounts(double[] productIdsAndAmounts)
     {
         var ingredients = new List<Ingredient>();
         var products = new List<Product>();
         
         for (int i = 0; i < productIdsAndAmounts.Length; i += 2)
         {
-            var productId = productIdsAndAmounts[i];
+            var productId = (int)productIdsAndAmounts[i];
             var amount = productIdsAndAmounts[i + 1];
             ingredients.Add(new Ingredient { ProductId = productId, AmountInGrams = amount });
             products.Add(ProductHelper.AllProducts.First(p => p.Id == productId));
